pyproject.toml
[tool.poetry]
name = "transactions-data"
version = "0.1.0"
description = ""
authors = ["Alik <alik220201@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pytest = "^8.3.3"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"


[tool.poetry.group.lint.dependencies]
isort = "^5.13.2"
black = "^24.10.0"
mypy = "^1.11.2"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
